# testdata/mux_crud/expected_openapi_legacy.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Generated API",
    "version": "1.0.0",
    "contact": {
      "name": "Anton Starikov",
      "url": "https://github.com/antst/go-apispec",
      "email": "antst@gmail.com"
    }
  },
  "paths": {
    "/orders": {
      "get": {
        "operationId": "testdata/mux_crud.listOrders",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/testdata_mux_crud.Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "testdata/mux_crud.createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/testdata_mux_crud.Order"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/testdata_mux_crud.Order"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain; charset=utf-8": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "operationId": "testdata/mux_crud.getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/testdata_mux_crud.Order"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "testdata/mux_crud.deleteOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "testdata_mux_crud.Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "total",
          "status"
        ]
      }
    }
  }
}